+++ tests/pest/Feature/SettingsTest.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * Tests the default plugin settings.
+ */
+
+use putyourlightson\blitz\Blitz;
+
+test('The cache control header doesn’t allow browser caching', function() {
+    expect(Blitz::$plugin->settings->cacheControlHeader)
+        ->toContain('max-age=0');
+});
+
+test('The expired cache control header doesn’t allow browser caching', function() {
+    expect(Blitz::$plugin->settings->cacheControlHeaderExpired)
+        ->toContain('max-age=0');
+});